w3c-reports/12092014/mac/runned-tests/cors/basic-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?get_value=hest_er_best&origin=none&1
Blocked access to external URL http://www1.localhost:8801/cors/resources/cors-makeheader.py?get_value=hest_er_best&origin=none&3
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8801/cors/resources/cors-makeheader.py?get_value=hest_er_best&origin=none&2. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.

Harness Error. harness_status.status = 2 , harness_status.message = null
